[passo3] mais metodos foram para o routes
