+++ vraptor-core/src/main/java/br/com/caelum/vraptor/http/route/Routes.java
@@ -4,14 +4,24 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.EnumSet;
 import java.util.List;
 
 import com.google.common.base.Predicate;
 import com.google.common.base.Supplier;
 import com.google.common.collect.FluentIterable;
 
-public class Routes {
+import br.com.caelum.vraptor.controller.HttpMethod;
 
+/*
+ * 6 pontos
+ * 
+ * depois de trazer mais métodos para operar aqui
+ * 
+ * 9 pontos
+ */
+public class Routes {
+	//2
 	private Collection<Route> routes = new PriorityRoutesList();
 	private static final Route NULL = new NoStrategy() {
 		@Override
@@ -28,6 +38,7 @@ public class Routes {
 		routes.add(r);
 	}
 	
+	//1 do if
 	public Collection<Route> routesMatchingUri(String uri) {
 		Collection<Route> routesMatchingURI = FluentIterable.from(routes)
 				.filter(canHandle(uri)).toSet();
@@ -38,6 +49,7 @@ public class Routes {
 		return routesMatchingURI;
 	}
 	
+	//1 do supplier
 	public Supplier<Route> lazyBla(final Class<?> rawtype, final Method method) {
 		return new Supplier<Route>() {
 			@Override
@@ -64,6 +76,37 @@ public class Routes {
 			@Override
 			public boolean apply(Route route) {
 				return route.canHandle(type, method);
+			}
+		};
+	}	
+	
+	// 1 ponto do if
+	public Collection<Route> routesMatchingUriAndMethod(String uri, HttpMethod method) {
+		Collection<Route> routesMatchingMethod = FluentIterable.from(routesMatchingUri(uri))
+				.filter(allow(method)).toSet();
+
+		if (routesMatchingMethod.isEmpty()) {
+			EnumSet<HttpMethod> allowed = allowedMethodsFor(uri);
+			throw new MethodNotAllowedException(allowed, method.toString());
+		}
+		return routesMatchingMethod;
+	}
+
+	//1 ponto do for
+	private EnumSet<HttpMethod> allowedMethodsFor(String uri) {
+		EnumSet<HttpMethod> allowed = EnumSet.noneOf(HttpMethod.class);
+		for (Route route : routesMatchingUri(uri)) {
+			allowed.addAll(route.allowedMethods());
+		}
+		return allowed;
+	}
+	
+	//1 ponto do predicate
+	private Predicate<Route> allow(final HttpMethod method) {
+		return new Predicate<Route>() {
+			@Override
+			public boolean apply(Route route) {
+				return route.allowedMethods().contains(method);
 			}
 		};
 	}
